requirements.txt
playwright==1.44.0
apscheduler==3.10.4
sentence-transformers==3.0.1
pdfplumber==0.11.0
python-dotenv==1.0.1
loguru==0.7.2
